requirements.txt
matplotlib==3.10.0
Pillow==11.1.0
torch==2.0.1
torchvision==0.20.1
transformers==4.47.1
numpy<2
ipywidgets>=7.6.0
jupyter>=1.0.0
pybind11>=2.12
scikit-image==0.25.0
tqdm==4.67.1